+++ zoomage/src/main/java/com/jsibbold/zoomage/OnScaleChangeListener.java
@@ -1,5 +1,8 @@
 package com.jsibbold.zoomage;
 
+/**
+ * Listener, that provides updates about scale changes in {@link ZoomageView}
+ */
 public interface OnScaleChangeListener {
     void onScaleChanged(float newScale);
 }
